+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Showing sync erroring users only on the info banner
+
 ### Security
 
 ## 3.1.0